Moving GIUH.hpp to dedicated subdirectory.
